Refactor embeddings datasets (#495)
