+++ src/eva/core/data/datasets/__init__.py
@@ -1,11 +1,11 @@
 """Datasets API."""
 
 from eva.core.data.datasets.base import Dataset
-from eva.core.data.datasets.dataset import TorchDataset
-from eva.core.data.datasets.embeddings import (
+from eva.core.data.datasets.classification import (
     EmbeddingsClassificationDataset,
     MultiEmbeddingsClassificationDataset,
 )
+from eva.core.data.datasets.dataset import TorchDataset
 
 __all__ = [
     "Dataset",
