+++ src/eva/core/data/datasets/classification/__init__.py
@@ -1,0 +1,8 @@
+"""Embedding cllassification datasets API."""
+
+from eva.core.data.datasets.classification.embeddings import EmbeddingsClassificationDataset
+from eva.core.data.datasets.classification.multi_embeddings import (
+    MultiEmbeddingsClassificationDataset,
+)
+
+__all__ = ["EmbeddingsClassificationDataset", "MultiEmbeddingsClassificationDataset"]
