--- tests/eva/core/data/datasets/embeddings/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for embeddings datasets."""
